Transition from make to cmake
This adds preliminary Windows support (compiles in VS2017). This is still a WIP and doesn't compile cleanly yet.

version.hpp was renamed to a more generic config.hpp.in (where config.hpp is generated by cmake). The version macros were changed to honor semver semantics as well as C++ guidelines (it is non-conformant to name things with leading underscores in C++; single-underscore is reserved for library vendors and double-underscore is reserved for compiler vendors).

It appears that the project owner likes using tabs, so added .editorconfig to enforce that preference.